requirements.txt
anthropic>=0.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
numpy>=1.26.0
pypdf>=4.0.0
python-docx>=1.1.0
